Made a change to the evtx module (bug fix).
Essentially if the evtx file is somehow broken and the underlying Parse::Evtx library cannot parse the file it will raise an Exception, which is caught by the tool, however the tool will return an empty timestamp object when this excception occurs. This on the other hand causes the tool to try to get the next event, that again returns an empty timestamp object, etc... making the tool go into an endless loop.

I added a counter, known bad counter, and check if it reaches 50 (increment it each time the exception is caught) I will return nothing back instead of an empty timestmap object. That way the tool will know that it should stop parsing the file, thus avoiding the potential endless loop while still trying to at least parse the file if it is corrupted.

Also changed the debian/control file (Tim - sentinelchicken) pointed out to me while he was isntalling the tool on his Debian/Testing machine it failed since not all dependencies were met. (the missing dependency was the libwww-perl library). Added that to the changelog

updated the timestamp of the Common file