Added a new feature to the tool, now it will allow -Z ZONE to be added.
-Z ZONE defines the timezone used in the output module.

before defining -z would actually define both the timezone of the host in question as well as the timezone of the output.
Now that can be separated, and the tool will print using another timezone than the hosts, for instance if you want the output to
be in GMT yet the host has EST5EDT timezone.