Updated Win.pm, to fix the problem with the known transform of timezones. That is a small bug was in the hash, so it was failing to compile/run.
Also updated a bit the hash, to add correct values and changed the pre-processor win_sysinfo
so that it now checks if a transform exists and compares the transformed time zone value to the
one chosen as a parameter to the tool. If they do not match the tool will update the timezone
accordingly.