Starting to upgrade the template to provide some assistance for the new style guide, providing proper documentation to the code.
Updated iehistory so it adheres to the new style guide. iehistory module
is to be the module everyone adores ;) at least in the perspective that it
will be the module that should adhere to the style guide and can be pointed
towards when people ask/query about the guide.

Updated slight license information in the mft.pm code to reflect some changes
to analyzeMFT licensing issues.