Created a perltidy configuration file for the project, based on the style guide.
I've also ran perltidy on all files in the project, and added considerable documentation to the main engine (Log2Timeline).

Need to complete the updated documentation for the main engine and the associated common libraries before moving on to individual
input modules.

This does not mean that all current files adhere perfectly to the style guide, but at least they have the look and feel, it is
just missing a code review and some of the rules that cannot be enforced by a perltidy type of script.