Updated the CHANGELOG to reflect some changes.
Updated the template_for_output_module to the new style guide and added some more
documentations to it.

Slight changes to l2t_process, that is ran perltidy against it.

Slight changes to Log2Timeline, still adding some documentations to it.

Minor changes to the csv module, documentation ones.

Added a new module, called serialize that instaed of outputting the data in
some preconfigured output format it simply uses JSON to serialie the output.

That way it is possible to read the output format and get back the original
hash value for the timestamp object (thus instill some further fine grained
filtering and sorting).