Updated CHANGELOG to reflect changes, also Common to reflect new version.
Makefile - Added a requirement for Test::More
Log2Timeline - Added a call to ->end if a verification fails.
Prefetch/Skype - small bugs that should hopefully fix that information is not fetched if recursive mode is used. (Bug reported by rlee).
t/* - Added few test files to the project, should automate testing.
That is this is the first indication that the project is getting TEST files, to do automatic testing, to at least minimize the changes of changes to the code making stuff less reliable.